Plot Heatmap Based on the Unique IDs Verified from the findUniqueIDs function

generate Heatmap Based on the Unique IDs matched to the netPhorce Data from findUniqueIDs() function

plotUniqueIDsHeatmaps(
  netPhorceData = netPhorceData,
  foundUniqueIDs = NULL,
  heatmapType = "Significant",
  minQVal = 0.05,
  plotly = TRUE
)

Arguments

netPhorceData

(Required). Processed netPhorce Object

foundUniqueIDs

(Required). Output from the findUniqueIDs function. If nothing in the “MatchedIDs” vector, no plot will be generated

heatmapType

(Required). Either “Significant” or “AbsencePresence” are required. “Significant” for the unique peptide IDs that were significant differentially phosphorylated peptides. “AbsencePresence” for the unique peptide IDs that were absent at at least one time point.

minQVal

(Required). A q-value threshold. A lower threshold increases stringency for displaying significant phosphopeptides into the heatmap. The q-value controls the positive false discovery rate and is estimated based on the p-values obtained from a linear mixed model fitted to the intensities of each peptide.

plotly

(Required). If TRUE, output an interactive plotly version, else output a static ggplot2 version.

Value

a ggplot/plotly object

Examples

if (FALSE) {
## Loading Two Conditions Example
data("twoConditionsExample")
## Identify the Key Columns
identifiedCols <- confirmColumnNames(rawMaxQuant = twoConditionsExample,
                                    positionCol = "Position",
                                    reverseCol = "Reverse",
                                    localizationProbCol = "Localization prob",
                                    potentialContaminationCol = "Potential contaminant",
                                    aminoAcidCol = "Amino acid",
                                    uniqueIDCol = "Protein",
                                    seqWindowIDCol = "Sequence window",
                                    fastaIDCol = "Fasta headers")
## Identify the Intensity Columns with Condition, Time Point and Replication Information
intensityCols <- confirmIntensityColumns(rawMaxQuant = twoConditionsExample,
                                         intensityPattern = "con_time_rep",
                                         verbose = TRUE)
## Process the data based on the identified columns
netPhorceData <- processData(rawMaxQuant = twoConditionsExample,
                             processedColNames = identifiedCols,
                             processedIntensity = intensityCols,
                             minReplication = 3,
                             minLocalProb = 0.75)
## Provided IDs all matched the IDs in the netPhorce Data
uniqueIDs <-
  findUniqueIDs(netPhorceData = netPhorceData,
                uniqueIDList = c(
                  # Significant Set Examples:
                  "AT1G13030.1", "AT1G13360.3", "AT1G42550.1",
                  # Unique/Abscence Set Examples
                  "AT1G17280.9", "AT1G22310.2", "AT1G23890.2"
                ),
                verbose = TRUE)
## Plotting Significant Heatmap - PLOTLY Version
plotUniqueIDsHeatmaps(netPhorceData = netPhorceData,
                      foundUniqueIDs = uniqueIDs,
                      heatmapType = "Significant",
                      minQVal = 0.1,
                      plotly = TRUE)
## Plotting Absence/Prescence Heatmap - GGPLOT version
plotUniqueIDsHeatmaps(netPhorceData = netPhorceData,
                      foundUniqueIDs = uniqueIDs,
                      heatmapType = "AbsencePresence",
                      plotly = FALSE)
}