Network Analysis

The phosphorylation signaling network shows the inferred causal relations between the kinases/phosphatases and downstream peptides. Network inference is based on dynamic Bayesian principles. For more than one condition, the common phosphopeptides between the conditions are used as input and a network is generated for each condition separately. As such, the network results might differ slightly when running each condition separately through NetPhorce compared to running the conditions together.

networkAnalysis(netPhorceData = netPhorceData, requestPlotData = TRUE)

Arguments

netPhorceData

(Required). Processed netPhorceData data

requestPlotData

(Required). If TRUE, data required to visualize the network in R via visNetwork package will be saved.

Value

The netPhorce object with the following slots:

Design

Contains data design information and filtering parameters

data.filtered

Contains all the data points in a Long format that passed filtering criteria.

data.filtered.aov.summary

Contains all the data points in a Long format with anova results.

Misc

Contains accessory data including default plotting colors and FASTA Keys, if present.

regulationData

Contains regulation data calculated through regulationCheck function

networkPhorceResutls

Contains the network results

networkPlotData

Contains data required to generate the network via visNetwork package

Examples

if (FALSE) {
## Loading One Condition Data
data("oneConditionExample")
## Identify the Key Columns
identifiedCols <- confirmColumnNames(rawMaxQuant = oneConditionExample,
                                    positionCol = "Position",
                                    reverseCol = "Reverse",
                                    localizationProbCol = "Localization prob",
                                    potentialContaminationCol = "Potential contaminant",
                                    aminoAcidCol = "Amino acid",
                                    uniqueIDCol = "Protein",
                                    seqWindowIDCol = "Sequence window",
                                    fastaIDCol = "Fasta headers")
## Identify the Intensity Columns with Condition, Time Point and Replication Information
intensityCols <- confirmIntensityColumns(rawMaxQuant = oneConditionExample,
                                         intensityPattern = "con_time_rep",
                                         verbose = TRUE)
## Process the data based on the identified columns
netPhorceData <- processData(rawMaxQuant = oneConditionExample,
                             processedColNames = identifiedCols,
                             processedIntensity = intensityCols,
                             minReplication = 3,
                             minLocalProb = 0.75)
## Validating the Kinase/Phosphatase Information
netPhorceData <- validateKinaseTable(netPhorceData = netPhorceData,
                                     defaultKinaseTable = TRUE,
                                     abbrev = "Ath")
## Regulation Validation based on user inputs
netPhorceData <- regulationCheck(netPhorceData = netPhorceData,
                                 upReg = 0.25,
                                 downReg = 0.25,
                                 absMinThreshold = 0.1,
                                 qValueCutOff = 0.05,
                                 verbose = TRUE)
## Network Analysis
networkData <- networkAnalysis(netPhorceData = netPhorceData,
                               requestPlotData = TRUE)
}